i think my 15% estimate was high

+++ src/main/java/com/robertx22/mine_and_slash/saveclasses/EntitySpellData.java
@@ -59,7 +59,7 @@ public class EntitySpellData {
         this.casterID = caster.getUniqueID()
             .toString();
 
-        this.lifeInTicks = config.duration;
+        this.lifeInTicks = (int) config.duration;
         this.configs = config;
     }
